linter fixes
- move snclientConfigFileWithScript function out of snclient_test.go to snclient_windows_test.go , since it is only being used to test windows powershell functions for now. it was not being used for other platforms
- remove some unused code lines
- use idiomatic assert functions
- line length
- nolint couple unwanted rules like unparam and duplicate
