+++ pkg/snclient/snclient_windows_test.go
@@ -8,7 +8,79 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
+
+// Generates a config file, where snclient can call a script.
+// scriptName does not have an extension
+// scriptFilename does have (most likely an OS specific) script extension.
+// It registers four commands for script
+// scriptName_arg1 : ./${SCRIPT_FILENAME} "$ARG1$"
+// scriptName_arg_numbered : ./${SCRIPT_FILENAME} "$ARG1$" "$ARG2$" "$ARG3$" "$ARG4$" "$ARG5$" "$ARG6$" "$ARG7$" "$ARG8$" "$ARG9$" "$ARG10$"
+// scriptName_args : ./${SCRIPT_FILENAME} "$ARGS$"
+// scriptName_args_quouted : ./${SCRIPT_FILENAME} "$ARGS"$"
+//
+//nolint:unparam // scriptName is so far always "powershell_detail" , no other test script uses this function. Keep it as a parameter for future use.
+func snclientConfigFileWithScript(t *testing.T, scriptsDir, scriptName, scriptFilename string) string {
+	t.Helper()
+
+	configTemplate := `
+[/modules]
+CheckExternalScripts = enabled
+
+[/paths]
+scripts = ${SCRIPTS_DIR}
+shared-path = %(scripts)
+
+[/settings/external scripts]
+timeout = 1111111
+allow arguments = true
+
+[/settings/external scripts/scripts]
+${SCRIPT_NAME}_arg1 = ./${SCRIPT_FILENAME} $ARG1$
+
+[/settings/external scripts/scripts/${SCRIPT_NAME}_arg1]
+allow arguments = true
+allow nasty characters = true
+
+[/settings/external scripts/scripts]
+${SCRIPT_NAME}_arg_numbered = ./${SCRIPT_FILENAME} $ARG1$ $ARG2$ $ARG3$ $ARG4$ $ARG5$ $ARG6$ $ARG7$ $ARG8$ $ARG9$ $ARG10$
+
+[/settings/external scripts/scripts/${SCRIPT_NAME}_arg_numbered]
+allow arguments = true
+allow nasty characters = true
+
+[/settings/external scripts/scripts]
+${SCRIPT_NAME}_args = ./${SCRIPT_FILENAME} $ARGS$
+
+[/settings/external scripts/scripts/${SCRIPT_NAME}_args]
+allow arguments = true
+allow nasty characters = true
+
+[/settings/external scripts/scripts]
+${SCRIPT_NAME}_args_quouted = ./${SCRIPT_FILENAME} $ARGS"$
+
+[/settings/external scripts/scripts/${SCRIPT_NAME}_args_quouted]
+allow arguments = true
+allow nasty characters = true
+`
+
+	mapper := func(placeholderName string) string {
+		switch placeholderName {
+		case "SCRIPTS_DIR":
+			return scriptsDir
+		case "SCRIPT_NAME":
+			return scriptName
+		case "SCRIPT_FILENAME":
+			return scriptFilename
+		default:
+			// if its not some value we know, leave it as is
+			return "$" + placeholderName
+		}
+	}
+
+	return os.Expand(configTemplate, mapper)
+}
 
 func TestMakeCmd(t *testing.T) {
 	config := ""
@@ -17,7 +89,7 @@ func TestMakeCmd(t *testing.T) {
 	commandString := `.\t\scripts\powershell_detail.ps1 -option1 option1 -option2 'option2' -option3 "option3" -option4 'option4.option4,option4:option4;option4|option4$option4' `
 	cmd, err := snc.makeCmd(context.TODO(), commandString)
 
-	assert.Nilf(t, err, "there should not be any errors when converting command: %s into an exec.Cmd of os/exec", commandString)
+	require.NoErrorf(t, err, "there should not be any errors when converting command: %s into an exec.Cmd of os/exec", commandString)
 
 	assert.NotEmptyf(t, cmd.SysProcAttr.CmdLine, "exec.Cmd from command: %s should not have an empty SysProcAttr.CmdLine", commandString)
 
@@ -42,7 +114,10 @@ func TestMakeCmd(t *testing.T) {
 	// users have to use it like this --optionX 'foo,bar' to have it accepted as a string
 
 	cmdLineExpectedContains := `-option1 option1 -option2 'option2' -option3 "option3" -option4 'option4.option4,option4:option4;option4|option4$option4'`
-	assert.Containsf(t, cmd.SysProcAttr.CmdLine, cmdLineExpectedContains, "exec.Cmd from command: %s\nshould contain this substring: %s\nbut it looks like this: %s", commandString, cmdLineExpectedContains, cmd.SysProcAttr.CmdLine)
+	assert.Containsf(t, cmd.SysProcAttr.CmdLine,
+		cmdLineExpectedContains,
+		"exec.Cmd from command: %s\nshould contain this substring: %s\nbut it looks like this: %s",
+		commandString, cmdLineExpectedContains, cmd.SysProcAttr.CmdLine)
 
 	var pathEnv string
 	for _, envVar := range cmd.Env {
@@ -51,13 +126,13 @@ func TestMakeCmd(t *testing.T) {
 		}
 	}
 
-	assert.NotEmpty(t, pathEnv, cmd.Env, "converted exec.Cmd from command: %s should contain PATH environment variable", commandString)
+	assert.NotEmpty(t, pathEnv, "converted exec.Cmd from command: %s should contain PATH environment variable", commandString)
 
 	// script is found under C:\Users\sorus\repositories\snclient\pkg\snclient\t
 	// scriptsPath, _ := snc.config.Section("/paths").GetString("scripts")
 	// assert.Containsf(t, pathEnv, scriptsPath+":", "converted exec.Cmd from command: %s should have its PATH variable: %s include the config ScriptsPath: %s", commandString, pathEnv, scriptsPath)
 
-	assert.Equal(t, true, cmd.SysProcAttr.HideWindow, "converted exec.Cmd from command: %s should hide its spawned window", commandString)
+	assert.Truef(t, cmd.SysProcAttr.HideWindow, "converted exec.Cmd from command: %s should hide its spawned window", commandString)
 
 	StopTestAgent(t, snc)
 }
@@ -75,7 +150,6 @@ func TestPowershell1(t *testing.T) {
 	res := snc.RunCheck("powershell_detail_arg1", []string{})
 
 	outputString := string(res.BuildPluginOutput())
-	// t.Logf("\n%s\n", outputString)
 
 	assert.Equalf(t, CheckExitOK, res.State, "check should return state ok")
 
@@ -88,7 +162,6 @@ func TestPowershell1(t *testing.T) {
 	for _, rawCommandLineExpectedItem := range rawCommandlineExpected {
 		assert.Containsf(t, outputString, rawCommandLineExpectedItem, "raw commandline should contain: %s", rawCommandLineExpectedItem)
 	}
-
 }
 
 func TestPowershellScriptArg1(t *testing.T) {
@@ -111,7 +184,6 @@ func TestPowershellScriptArg1(t *testing.T) {
 	assert.NotNilf(t, checkData, "check should return a checkData")
 
 	outputString := string(checkResult.BuildPluginOutput())
-	// t.Logf("\n%s\n", outputString)
 
 	assert.Equalf(t, CheckExitOK, checkResult.State, "check should return state OK")
 
@@ -138,6 +210,7 @@ func TestPowershellScriptArg1(t *testing.T) {
 	}
 }
 
+//nolint:dupl // the functions are largely the same, but scriptMacroType is different. Redefining expected strings for each macro type is easier to understand.
 func TestPowershellScriptArgNumbered(t *testing.T) {
 	testDir, _ := os.Getwd()
 	scriptsDir := filepath.Join(testDir, "t", "scripts")
@@ -167,7 +240,6 @@ func TestPowershellScriptArgNumbered(t *testing.T) {
 	assert.NotNilf(t, checkData, "check should return a checkData")
 
 	outputString := string(checkResult.BuildPluginOutput())
-	//t.Logf("\n%s\n", outputString)
 
 	assert.Equalf(t, CheckExitOK, checkResult.State, "check should return state OK")
 
@@ -191,6 +263,7 @@ func TestPowershellScriptArgNumbered(t *testing.T) {
 	}
 }
 
+//nolint:dupl // the functions are largely the same, but scriptMacroType is different. Redefining expected strings for each macro type is easier to understand.
 func TestPowershellScriptArgs(t *testing.T) {
 	testDir, _ := os.Getwd()
 	scriptsDir := filepath.Join(testDir, "t", "scripts")
@@ -220,7 +293,6 @@ func TestPowershellScriptArgs(t *testing.T) {
 	assert.NotNilf(t, checkData, "check should return a checkData")
 
 	outputString := string(checkResult.BuildPluginOutput())
-	//t.Logf("\n%s\n", outputString)
 
 	assert.Equalf(t, CheckExitOK, checkResult.State, "check should return state OK")
